Ditch the androidx.preferences library
Coding things by myself is more boilerplaty, but it is much more
straightforward. That library was a mess.

One of the motivations is that it was that some things are not supported
by this library. In particular, combining a timepicker with a toggle,
and and displaying additional information next to the preferences.