Bugfix: make reminders last the whole day
They were originally intended to last 14 hours, but due to a bug only
lasted 14 minutes. This meant that most of the time, if you pressed the
"not now" button in the morning the reminder would not show up
afterwards, since it had already expired by that point.

Anyway, to make the notification behavior more consistent with the rest
of the app, I'm now having it last until the end of the day instead of
using a fixed timeout.